+++ pdf/lib/pdf.dart
@@ -24,6 +24,7 @@ export 'src/pdf/font/font_metrics.dart';
 export 'src/pdf/font/ttf_parser.dart';
 export 'src/pdf/format/name.dart';
 export 'src/pdf/format/object_base.dart' show DeflateCallback, PdfVersion;
+export 'src/pdf/format/stream.dart';
 export 'src/pdf/graphic_state.dart';
 export 'src/pdf/graphics.dart';
 export 'src/pdf/obj/annotation.dart';